Organize code.
Moved the the MyOffers action to the OffersController and moved the related view models and view to the Offers folder. They should belong to the Offers folder since the action has to do with offers.
