+++ OrangeBricks.Web/Controllers/Property/PropertyController.cs
@@ -74,15 +74,6 @@ namespace OrangeBricks.Web.Controllers.Property
         }
 
         [OrangeBricksAuthorize(Roles = "Buyer")]
-        public ActionResult MyOffers()
-        {
-            var builder = new MyOffersViewModelBuilder(_context);
-            var viewModel = builder.Build(User.Identity.GetUserId());
-
-            return View(viewModel);
-        }
-
-        [OrangeBricksAuthorize(Roles = "Buyer")]
         public ActionResult MakeOffer(int id)
         {
             var builder = new MakeOfferViewModelBuilder(_context);
