+++ OrangeBricks.Web/OrangeBricks.Web.csproj
@@ -186,9 +186,9 @@
     <Compile Include="Controllers\Property\PropertyController.cs" />
     <Compile Include="Controllers\Property\ViewModels\CreatePropertyViewModel.cs" />
     <Compile Include="Controllers\Property\ViewModels\MakeOfferViewModel.cs" />
-    <Compile Include="Controllers\Property\ViewModels\MyOffersViewModel.cs" />
+    <Compile Include="Controllers\Offers\ViewModels\MyOffersViewModel.cs" />
     <Compile Include="Controllers\Property\ViewModels\MyPropertiesViewModel.cs" />
-    <Compile Include="Controllers\Property\ViewModels\OfferViewModel.cs" />
+    <Compile Include="Controllers\Offers\ViewModels\MyOfferViewModel.cs" />
     <Compile Include="Controllers\Property\ViewModels\PropertiesViewModel.cs" />
     <Compile Include="Controllers\Property\ViewModels\PropertyViewModel.cs" />
     <Compile Include="Global.asax.cs">
@@ -291,7 +291,7 @@
     <Content Include="Views\Property\MakeOffer.cshtml" />
     <Content Include="Views\Offers\OnProperty.cshtml" />
     <None Include="Views\Account\UnAuthorized.cshtml" />
-    <Content Include="Views\Property\MyOffers.cshtml" />
+    <Content Include="Views\Offers\MyOffers.cshtml" />
   </ItemGroup>
   <ItemGroup>
     <Folder Include="App_Data\" />
