+++ OrangeBricks.Web/Views/Shared/_Layout.cshtml
@@ -32,7 +32,7 @@
                         }
                         if (User.IsInRole("Buyer"))
                         {
-                            <li>@Html.ActionLink("My Offers", "MyOffers", "Property")</li>
+                            <li>@Html.ActionLink("My Offers", "MyOffers", "Offers")</li>
                         }
                     }
                 </ul>
